clear App component with legacy code + work on lapikit website
