+++ apps/website/src/lib/components/footer.svelte
@@ -1,8 +1,28 @@
 <script lang="ts">
-	let { children, accessOnRGPD = $bindable(), ...rest } = $props();
+	import { t } from '$lib/i18n';
+	import { Btn, Separator } from 'lapikit/components';
+	import type { Snippet } from 'svelte';
+
+	let {
+		children,
+		accessOnRGPD = $bindable(),
+		...rest
+	}: { children?: Snippet; accessOnRGPD?: boolean } = $props();
+
+	let year = new Date().getFullYear();
 </script>
 
 <footer {...rest}>
 	{@render children?.()}
-	<button onclick={() => (accessOnRGPD = true)}>cookie settings</button>
+
+	<div class="flex items-center justify-between">
+		<p>©{year} Lapikit.</p>
+		<div class="flex items-center gap-2">
+			<Btn href="https://nycolaide.dev" target="_blank" size="sm" variant="text">Nycolaide</Btn>
+			<Separator orientation="vertical" />
+			<Btn variant="text" size="sm" onclick={() => (accessOnRGPD = true)}>
+				{$t('app.cookie_settings')}
+			</Btn>
+		</div>
+	</div>
 </footer>
