+++ packages/lapikit/src/components/app/app.svelte
@@ -11,8 +11,6 @@
 	});
 </script>
 
-{$modalOpen ? ($modalOpen === 'persistent' ? 'persistent' : 'true') : 'false'}
-
 {@render children?.()}
 
 <!-- svelte-ignore a11y_no_static_element_interactions -->
